bot_candidates/formula-trend/05/requirements.txt
pillow
eventlet
flask-socketio==1.0b1
numpy==1.14.5
opencv-python==3.4.2.17
tensorflow==1.10.1
keras==2.2.2
pandas==0.23.4
scipy==1.1.0
h5py==2.8.0
